+++ changelog.md
@@ -1,6 +1,8 @@
 ## 0.5.0
 
  - Base errors support added (i.e. errors flow through all kind of transformations/combinations)
+ - New method `.onError`
+ - New method `.offError`
  - Emitter now has `.error` method for emitting errors
  - Emitter object also now has `.error` method
  - Properties now may have a current error (as well as current value)